No longer duplicate the filename in the policy documents
